Update the Jupyter UI for 0.5 (#2848)
* Add polling logic

* Get pvcs
* Return more info in Notebooks
* Update Makefile, update css/html
* Update Rok UI
* Minor fixes on html and flask debug mode
* Update add-notebook.js in Rok
* Add html for new notebook in ROK
* Make UI to poll the Backend for State


* Update the GCB script


* Organize file structure and refactor imports


* Sync Notebook name with Workspace Volume


* Use different warning if no DefStorClass


* Make backend REST


* Change Notebook status


* Misc changes


* Scroll on top to show the error


* Clean notebook table


* Add dialogs

If a user wants to delete a notebook, provide a confirm dialog


* Use term Namespace instead of Project


* Fix CSS rendering


* Make the polling exponential

The frontend will stop polling the backend exponentially.
If the state doesn't change, reduce the retries counter.

If the retries counter goes to zero, then double the time between
requests and reset the retries counter. Maximum poll time is 16 seconds.

If the state changes, or the user creates/deletes a Notebook or
selects a different namespace; reset the polling mechanism


* Handle no Status in Notebook CR
