--- components/jupyter-web-app/baseui/.gitignore
@@ -1,1 +1,0 @@
-/__pycache__
