--- components/jupyter-web-app/default/kubeflow/jupyterui/.gitignore
@@ -1,1 +1,0 @@
-/__pycache__
