--- components/jupyter-web-app/default/kubeflow/jupyterui/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask, url_for, jsonify
-
-app = Flask(__name__)
-
-from kubeflow.jupyterui import routes
