--- components/jupyter-web-app/default/kubeflow/jupyterui/templates/notebooks.html
@@ -1,79 +1,0 @@
-{% extends "layout.html" %}
-{% block javascript %}
-  {{ super() }}
-  <script src="{{ prefix + url_for('static', filename='js/notebooks.js') }}">
-  </script>
-{% endblock %}
-
-{% block content %}
-<!-- Namespaces Control Buttons -->
-<table class="ns" align="center">
-  <tr>
-    <td>
-
-    </td>
-    <td>
-      <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
-        <select class="mdl-textfield__input" id="ns-select" name="ns">
-          <!-- <option value="default">default</option> -->
-          <!-- Fill with the namespaces the user can see -->
-          {% for ns in namespaces %}
-            <option value="{{ ns }}">{{ ns }}</option>
-          {% endfor %}
-        </select>
-        <label class="mdl-textfield__label" for="ns">Namespace</label>
-      </div>
-    </td>
-    <td>
-      <div class="mdl-textfield mdl-js-textfield mdl-textfield--expandable">
-        <label class="mdl-button mdl-js-button mdl-button--icon" for="search-nb">
-          <i class="material-icons">search</i>
-        </label>
-        <div class="mdl-textfield__expandable-holder">
-          <input class="mdl-textfield__input" type="text" id="search-nb" >
-          <label class="mdl-textfield__label" for="search-nb">Enter your query...</label>
-        </div>
-      </div>
-    </td>
-  </tr>
-</table>
-
-<!-- The Notebooks Table -->
-<div class="mdl-grid center-items">
-  <div class="mdl-cell mdl-cell--8-col">
-    <!-- Table Card -->
-    <div class="demo-card-wide mdl-card mdl-shadow--3dp wide">
-
-      <div class="mdl-card__title mdl-card--table__header">
-        <h2 class="mdl-card__title-text mdl-typography--title">Notebooks</h2>
-        <button class="mdl-button mdl-button--icon mdl-js-button mdl-js-ripple-effect"
-                id="create-nb-btn">
-          <i class="material-icons">add</i>
-        </button>
-      </div>
-
-      <div class="mdl-card__table">
-        <table class="mdl-data-table mdl-js-data-table wide">
-          <thead>
-            <tr>
-              <th class="mdl-data-table__cell--non-numeric">Name</th>
-              <th class="">Namespaces</th>
-              <th>Actions</th>
-            </tr>
-          </thead>
-          <tbody id="nb-table-body">
-            <!-- Dynamically load the page -->
-          </tbody>
-        </table>
-      </div>
-
-      <!-- Menus on top right of top card -->
-      <div class="mdl-card__menu">
-        <button class="mdl-button mdl-button--icon mdl-js-button mdl-js-ripple-effect">
-          <i class="material-icons">filter_list</i>
-        </button>
-      </div>
-    </div>
-  </div>
-</div>
-{% endblock content %}
