+++ components/jupyter-web-app/kubeflow_jupyter/default/static/kubeflow.png
