+++ components/jupyter-web-app/kubeflow_jupyter/rok/static/kubeflow.png
