+++ components/jupyter-web-app/kubeflow_jupyter/rok/templates/layout.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <!-- Required meta tags -->
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+  <!-- Material Design Lite -->
+  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+  <link rel="stylesheet" href="https://code.getmdl.io/1.3.0/material.blue-pink.min.css">
+  <script src="https://code.getmdl.io/1.3.0/material.min.js"></script>
+
+  <!-- UI Specific -->
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+  <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css">
+  <link rel="shortcut icon" href="{{ prefix + url_for('static', filename='kubeflow.png') }}">
+  <link rel="stylesheet" href="{{ prefix + url_for('static', filename='css/main.css') }}">
+
+  <!-- Dialogs -->
+  <link rel="stylesheet" href="{{ prefix + url_for('static', filename='css/dialog.css') }}"">
+  <script src="{{ prefix + url_for('static', filename='js/dialog.js') }}"></script>
+  
+  {% if title %}
+      <title>Kubeflow Jupyter - {{ title }}</title>
+  {% else %}
+      <title>Kubeflow Jupyter</title>
+  {% endif %}
+
+  {% block javascript %}
+    <script type="text/javascript">
+      var prefix = "{{ prefix }}";
+    </script>
+  {% endblock %}
+</head>
+<body>
+  <div class='parent errors'>
+    <div class="spacer"></div>
+    <div id="error-msgs"></div>
+    <div class="spacer"></div>
+  </div>
+  <div class="page-content">
+    {% block content %}{% endblock %}
+  </div>
+</body>
+</html>
