--- components/jupyter-web-app/rok/kubeflow/rokui/.gitignore
@@ -1,1 +1,0 @@
-/__pycache__
