--- components/jupyter-web-app/rok/kubeflow/rokui/static/css/main.css
@@ -1,116 +1,0 @@
-.wide {
-    width: 100%;
-}
-
-.no-ud-pad {
-    padding-top: 0px;
-    padding-bottom: 0px;
-}
-
-.nopad {
-    padding: 0 0 0 0;
-    margin: 0 0 8px;
-}
-
-.white {
-    color: #fff;
-}
-
-.down {
-    padding-bottom: 10px;
-}
-
-.hidden {
-    display: none;
-}
-
-.my-table {
-    width: 75%;
-    top: 10px;
-}
-
-.mdl-textfield__label{
-   margin-bottom:2px;
- }
-
-/*---------- Warnings ---------------*/
-
-.alert-warning {
-    color: #8a6d3b;
-    background-color: #fcf8e3;
-    border-color: #faebcc;
-}
-
-.alert-danger {
-    color: #a94442;
-    background-color: #f2dede;
-    border-color: #ebccd1;
-}
-
-.alert-success {
-    color: #155724;
-    background-color: #d4edda;
-    border-color: #c3e6cb;
-}
-
-.alert {
-    padding: 15px;
-    margin-bottom: 20px;
-    border-radius: 4px;
-}
-
-.close {
-  margin-left: 15px;
-  font-weight: bold;
-  float: right;
-  font-size: 22px;
-  line-height: 20px;
-  cursor: pointer;
-  transition: 0.3s;
-}
-
-.closebtn:hover {
-  color: black;
-}
-
-/*---------- Card Table -------------*/
-
-.demo-card-wide.mdl-card {
-  overflow: visible !important;
-  min-height: 0;
-  min-width: 0;
-}
-
-.fab, .fas, .far {
-  padding-left: 5px;
-  padding-right: 5px;
-}
-
-
-/*---------- Login -------------*/
-.mdl-grid.center-items {
-  justify-content: center;
-}
-
-.card-cont {
-    padding-top: 15px;
-    padding-left: 15px;
-    padding-right: 15px;
-    padding-bottom: 15px;
-}
-
-.card-header {
-    background:#2196f3;
-    padding: 0 0 0 0; /* Normal sized borders */
-}
-
-.mdl-card__supporting-text {
-    padding: 0 0 0 0;
-    width: 100%;
-    overflow: visible;
-}
-
-.mdl-card {
-  overflow: visible;
-  z-index: 100;
-}
