--- components/jupyter-web-app/rok/run.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-from kubeflow.rokui import app
-
-if __name__ == '__main__':
-  app.run(debug=True, host="0.0.0.0")
